A Hangman game implemented using HTML, CSS (Bootstrap), and JavaScript. The game allows players to guess words by entering letters. The HTML structure consists of a header, navigation bar, main content area, and input section. The JavaScript code handles the logic of the game, including word selection, hints, scoring, level progression, and checking user input
The HTML code begins with the standard document structure, including meta tags for defining the page's properties. The page title is set as "Hangman . Level one." The CSS Bootstrap framework is linked for styling the page. The body contains a header section with a scrolling marquee displaying the game title, and a red-colored hint box for providing hints to the player.

The navigation section displays information about the game, such as the category, level, and score. The "completed" section remains hidden initially but is displayed when the player completes Level 1. The main section of the page includes a container with images representing the different stages of the hangman. Another container displays a hidden paragraph that will show the correct word if the player loses.

The JavaScript code begins by defining an array of words that can be guessed in the game. It also initializes variables and selects a random word from the array to begin the game. The "select" function chooses a word based on a random index and displays a corresponding hint based on the selected word's category.

The game keeps track of the player's score and level. The score increments by 20 for each correct guess, and the level increases as the player progresses. When the score reaches 100 and the level is 6, Level 1 is completed, and a congratulatory message is displayed. The information section is hidden, and a "proceed" button appears, allowing the player to proceed to the next stage. However, the functionality for transitioning to the next stage is currently commented out.

The "checks" function handles the logic for checking user input. If the input is empty, an error message is displayed. If the input matches the selected word, the score and level increase, and a new word is selected. If the input is incorrect, the hangman's body parts are gradually displayed, and after six incorrect attempts, the game ends with a "GAME OVER" message and displays the correct word.